app/requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4
alembic==1.6.5
fastapi==0.63.0
httpx==0.18.2
Jinja2==2.11.2
phonenumbers==8.12.17
psycopg2==2.9.1
pytest-asyncio==0.15.1
requests==2.25.1
sirius-sdk==1.2.2
supervisor==4.2.0
sqlalchemy==1.4.19
uvicorn>=0.11.8
websockets==8.1
